split source files into files
