+++ pokerSingle/pokerSingle.vcxproj
@@ -71,15 +71,23 @@
   <PropertyGroup Label="UserMacros" />
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">
     <LinkIncremental>true</LinkIncremental>
+    <OutDir>$(SolutionDir)\bin\$(Platform)\$(Configuration)\</OutDir>
+    <IntDir>$(SolutionDir)\bin\intermediates\$(Platform)\$(Configuration)\</IntDir>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">
     <LinkIncremental>true</LinkIncremental>
+    <OutDir>$(SolutionDir)\bin\$(Platform)\$(Configuration)\</OutDir>
+    <IntDir>$(SolutionDir)\bin\intermediates\$(Platform)\$(Configuration)\</IntDir>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">
     <LinkIncremental>false</LinkIncremental>
+    <OutDir>$(SolutionDir)\bin\$(Platform)\$(Configuration)\</OutDir>
+    <IntDir>$(SolutionDir)\bin\intermediates\$(Platform)\$(Configuration)\</IntDir>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|x64'">
     <LinkIncremental>false</LinkIncremental>
+    <OutDir>$(SolutionDir)\bin\$(Platform)\$(Configuration)\</OutDir>
+    <IntDir>$(SolutionDir)\bin\intermediates\$(Platform)\$(Configuration)\</IntDir>
   </PropertyGroup>
   <ItemDefinitionGroup Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">
     <ClCompile>
@@ -143,42 +151,42 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClInclude Include="Bot.h" />
-    <ClInclude Include="Card.h" />
-    <ClInclude Include="ConsolePlayer.h" />
-    <ClInclude Include="CustomExceptions.h" />
-    <ClInclude Include="Decision.h" />
-    <ClInclude Include="Deck.h" />
-    <ClInclude Include="Flags.h" />
-    <ClInclude Include="Functions.h" />
-    <ClInclude Include="Game.h" />
-    <ClInclude Include="Hand.h" />
-    <ClInclude Include="HandAttempt.h" />
-    <ClInclude Include="MemoryTracking.h" />
-    <ClInclude Include="Player.h" />
-    <ClInclude Include="PotsHandler.h" />
-    <ClInclude Include="random.h" />
-    <ClInclude Include="Pot.h" />
-    <ClInclude Include="sorted_vector.h" />
-    <ClInclude Include="Table.h" />
+    <ClInclude Include="src\players\Bot.h" />
+    <ClInclude Include="src\card_related_items\Card.h" />
+    <ClInclude Include="src\players\ConsolePlayer.h" />
+    <ClInclude Include="src\tools\CustomExceptions.h" />
+    <ClInclude Include="src\game_related_items\Decision.h" />
+    <ClInclude Include="src\card_related_items\Deck.h" />
+    <ClInclude Include="src\Flags.h" />
+    <ClInclude Include="src\tools\Functions.h" />
+    <ClInclude Include="src\game_related_items\Game.h" />
+    <ClInclude Include="src\card_related_items\Hand.h" />
+    <ClInclude Include="src\card_related_items\HandAttempt.h" />
+    <ClInclude Include="src\tools\MemoryTracking.h" />
+    <ClInclude Include="src\game_related_items\Player.h" />
+    <ClInclude Include="src\game_related_items\PotsHandler.h" />
+    <ClInclude Include="src\tools\random.h" />
+    <ClInclude Include="src\game_related_items\Pot.h" />
+    <ClInclude Include="src\tools\sorted_vector.h" />
+    <ClInclude Include="src\card_related_items\Table.h" />
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="Bot.cpp" />
-    <ClCompile Include="Card.cpp" />
-    <ClCompile Include="ConsolePlayer.cpp" />
-    <ClCompile Include="Decision.cpp" />
-    <ClCompile Include="Deck.cpp" />
-    <ClCompile Include="Functions.cpp" />
-    <ClCompile Include="Game.cpp" />
-    <ClCompile Include="Hand.cpp" />
-    <ClCompile Include="HandAttempt.cpp" />
-    <ClCompile Include="MemoryTracking.cpp" />
-    <ClCompile Include="Player.cpp" />
-    <ClCompile Include="pokerSingle.cpp" />
-    <ClCompile Include="PotsHandler.cpp" />
-    <ClCompile Include="random.cpp" />
-    <ClCompile Include="Pot.cpp" />
-    <ClCompile Include="Table.cpp" />
+    <ClCompile Include="src\players\Bot.cpp" />
+    <ClCompile Include="src\card_related_items\Card.cpp" />
+    <ClCompile Include="src\players\ConsolePlayer.cpp" />
+    <ClCompile Include="src\game_related_items\Decision.cpp" />
+    <ClCompile Include="src\card_related_items\Deck.cpp" />
+    <ClCompile Include="src\tools\Functions.cpp" />
+    <ClCompile Include="src\game_related_items\Game.cpp" />
+    <ClCompile Include="src\card_related_items\Hand.cpp" />
+    <ClCompile Include="src\card_related_items\HandAttempt.cpp" />
+    <ClCompile Include="src\tools\MemoryTracking.cpp" />
+    <ClCompile Include="src\game_related_items\Player.cpp" />
+    <ClCompile Include="src\main.cpp" />
+    <ClCompile Include="src\game_related_items\PotsHandler.cpp" />
+    <ClCompile Include="src\tools\random.cpp" />
+    <ClCompile Include="src\game_related_items\Pot.cpp" />
+    <ClCompile Include="src\card_related_items\Table.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="pokerSingle.rc" />
