+++ pokerSingle/src/tools/Functions.cpp
