+++ pokerSingle/src/tools/Functions.h
@@ -2,7 +2,7 @@
 #ifndef FUNCTIONS_H_
 #define FUNCTIONS_H_
 
-#include "Flags.h"
+#include "../Flags.h"
 #include <iostream>
 
 /**
