+++ pokerSingle/src/tools/random.cpp
